# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MDDB API
  description: |
    MDDB (Markdown Database) is a high-performance, version-controlled markdown database with triple protocol support.

    ## Features
    - Store and manage markdown documents with metadata
    - Full revision history for all documents
    - Multi-language support
    - Fast metadata-based search
    - **Vector search** with embedding providers (OpenAI, Cohere, Voyage AI, Ollama)
    - **Full-text search** with TF-IDF, BM25, and BM25F field-weighted scoring
    - **Hybrid search** combining FTS + vector search with alpha blending or RRF fusion
    - **Webhooks** for real-time event notifications
    - **TTL** (time-to-live) support for automatic document expiration
    - **JSON Schema validation** for document metadata
    - **URL import** with YAML frontmatter parsing
    - **Authentication** - JWT tokens and API keys
    - **Authorization** - Collection-level RBAC (Read/Write/Admin)
    - **User Management** - Multi-user with admin roles
    - **Group Permissions** - Organize users into groups
    - Template variable substitution
    - Export capabilities (NDJSON, ZIP)
    - Backup and restore operations
    - **Replication** - Leader-follower replication with binlog streaming

    ## Protocols
    - **HTTP/JSON** - Port 11023 (documented here)
    - **gRPC/Protobuf** - Port 11024 (see gRPC documentation)
    - **GraphQL** - Port 11023/graphql (enable with MDDB_GRAPHQL_ENABLED=true)
...

    ## Authentication
...

...

...
    ```

...
servers:
...
paths:
...
  /v1/add:
    post:
      tags:
        - Documents
      summary: Add or update document
      description: |
        Add a new document or update an existing one. If a document with the same collection, key, and language exists, it will be updated and a new revision will be created.
      operationId: addDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRequest"
      responses:
        "200":
          description: Document added/updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/get:
    post:
      tags:
        - Documents
      summary: Get document
      description: |
        Retrieve a document by collection, key, and language. Supports template variable substitution using the `env` parameter.
      operationId: getDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetRequest"
      responses:
        "200":
          description: Document retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/search:
    post:
      tags:
        - Search
      summary: Search documents
      description: |
        Search documents with metadata filtering, sorting, and pagination.

        **Filtering:** AND logic between different metadata keys, OR logic between values of the same key.

        **Sorting:** Sort by `addedAt`, `updatedAt`, or `key`.
      operationId: searchDocuments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          headers:
            X-Total-Count:
              description: Total number of matching documents (before pagination)
              schema:
                type: integer
                example: 234
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/delete:
    post:
      tags:
        - Documents
      summary: Delete document
      description: Delete a specific document by collection, key, and language
      operationId: deleteDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
      responses:
        "200":
          description: Document deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/delete-collection:
    post:
      tags:
        - Documents
      summary: Delete collection
      description: Delete all documents in a collection
      operationId: deleteCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCollectionRequest"
      responses:
        "200":
          description: Collection deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    description: Number of documents deleted
                    example: 42
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/export:
    post:
      tags:
        - Export
      summary: Export documents
      description: |
        Export documents in NDJSON or ZIP format with optional metadata filtering.

        **NDJSON:** Newline-delimited JSON, one document per line.

        **ZIP:** Archive with markdown files named `{key}.{lang}.md`.
      operationId: exportDocuments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: Export file
          content:
            application/x-ndjson:
              schema:
                type: string
                format: binary
            application/zip:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/backup:
    get:
      tags:
        - Export
      summary: Create backup
      description: Create a backup of the database file
      operationId: createBackup
      parameters:
        - name: to
          in: query
          required: true
          description: Destination path for backup file
          schema:
            type: string
            example: backup-20250109.db
      responses:
        "200":
          description: Backup created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  backup:
                    type: string
                    example: backup-20250109.db
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/restore:
    post:
      tags:
        - Export
      summary: Restore from backup
      description: Restore database from a backup file
      operationId: restoreBackup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
              properties:
                from:
                  type: string
                  description: Path to backup file
                  example: backup-20250109.db
      responses:
        "200":
          description: Database restored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  restored:
                    type: string
                    example: backup-20250109.db
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/truncate:
    post:
      tags:
        - Maintenance
      summary: Truncate revision history
      description: |
        Remove old revisions to save disk space. Keeps the latest N revisions per document.

        **Warning:** This operation is irreversible.
      operationId: truncateRevisions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TruncateRequest"
      responses:
        "200":
          description: Revisions truncated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    description: Number of revisions deleted
                    example: 150
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/vector-search:
    post:
      tags:
        - Vector Search
      summary: Semantic vector search
      description: |
        Search documents by semantic similarity using vector embeddings.

        Provide either a `query` string (which will be embedded automatically using the configured provider) or a raw `queryVector` array of floats.

        Results are ranked by cosine similarity score. Use `threshold` to filter out low-relevance results. Use `filterMeta` to narrow results by metadata before ranking.

        Use `algorithm` to select the search strategy: `flat` (exact), `hnsw` (fast approximate), `ivf` (clustered), or `pq` (compressed). See [Search Algorithms](SEARCH.md) for details.
      operationId: vectorSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VectorSearchRequest"
      responses:
        "200":
          description: Vector search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VectorSearchResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Vector index is still loading
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/vector-reindex:
    post:
      tags:
        - Vector Search
      summary: Reindex collection embeddings
      description: |
        Trigger re-embedding of all documents in a collection. This is a write-protected endpoint.

        Use `force: true` to re-embed all documents even if their content has not changed since the last embedding.
      operationId: vectorReindex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VectorReindexRequest"
      responses:
        "200":
          description: Reindex completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VectorReindexResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/fts:
    post:
      tags:
        - Full-Text Search
      summary: Full-text search
      description: |
        Perform keyword-based full-text search across documents in a collection using TF-IDF, BM25, or BM25F scoring.

        The query is tokenized and matched against document content. Results include matched terms and relevance scores.

        **Algorithms:**
        - `tfidf` (default): Classic TF-IDF scoring
        - `bm25`: Okapi BM25 with document length normalization (k1=1.2, b=0.75)
        - `bm25f`: BM25F field-weighted scoring — applies different weights to title, metadata, and content fields

        **In-graph filtering:** Use `filterMeta` to narrow results by metadata before scoring, same as vector search.
      operationId: fullTextSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FTSSearchRequest"
      responses:
        "200":
          description: Full-text search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FTSSearchResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/hybrid-search:
    post:
      tags:
        - Hybrid Search
      summary: Hybrid sparse+dense search
      description: |
        Combine FTS (keyword) and vector (semantic) search into a single query. Returns results ranked by a fused score.

        **Fusion strategies:**
        - `alpha` (default): Weighted alpha blending — `combined = (1-alpha) * normalizedFTS + alpha * vectorScore`. Alpha=0 is pure keyword, alpha=1 is pure semantic.
        - `rrf`: Reciprocal Rank Fusion — `score = 1/(k + rank_fts) + 1/(k + rank_vector)`. Rank-based, no alpha needed.

        **FTS algorithms:** `bm25` (default), `bm25f` (field-weighted).

        **Vector algorithms:** `flat` (default), `hnsw`, `ivf`, `pq`, `sq`.

        Use `filterMeta` to pre-filter by metadata before both FTS and vector scoring.
      operationId: hybridSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HybridSearchRequest"
      responses:
        "200":
          description: Hybrid search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HybridSearchResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/import-url:
    post:
      tags:
        - Import
      summary: Import document from URL
      description: |
        Fetch content from a URL, parse YAML frontmatter if present, merge metadata, and store the document. This is a write-protected endpoint.

        If `key` is not provided, it is derived from the URL path. YAML frontmatter metadata from the fetched content is merged with the `meta` parameter.
      operationId: importUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportURLRequest"
      responses:
        "200":
          description: Document imported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/set-ttl:
    post:
      tags:
        - TTL
      summary: Set document TTL
      description: |
        Set a time-to-live (TTL) on a document. After the TTL expires, the document will be automatically deleted. This is a write-protected endpoint.

        Set `ttl` to `0` to remove the TTL and make the document permanent again.
      operationId: setTTL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetTTLRequest"
      responses:
        "200":
          description: TTL set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/webhooks:
    get:
      tags:
        - Webhooks
      summary: List webhooks
      description: List all registered webhooks.
      operationId: listWebhooks
      responses:
        "200":
          description: List of registered webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      tags:
        - Webhooks
      summary: Register webhook
      description: |
        Register a new webhook to receive event notifications. This is a write-protected endpoint.

        If `collection` is empty, the webhook will receive events for all collections.
      operationId: registerWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWebhookRequest"
      responses:
        "200":
          description: Webhook registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/webhooks/delete:
    post:
      tags:
        - Webhooks
      summary: Delete webhook
      description: Delete a registered webhook by its ID. This is a write-protected endpoint.
      operationId: deleteWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteWebhookRequest"
      responses:
        "200":
          description: Webhook deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
                  id:
                    type: string
                    example: abc123
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/schema/set:
    post:
      tags:
        - Schema Validation
      summary: Set collection schema
      description: |
        Set a JSON Schema for a collection's metadata validation. This is a write-protected endpoint.

        Once set, all documents added to the collection will have their metadata validated against this schema.
      operationId: setSchema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSchemaRequest"
      responses:
        "200":
          description: Schema set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  collection:
                    type: string
                    example: blog
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Read-only mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: document not found
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /v1/auth/login
    APIKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from POST /v1/auth/api-key